refactor: Renames files and form to match latest updates
Renames view file and form name to match the latest updates in the remediation json response
