+++ playground/mocks/data/idp/idx/authenticator-expired-custom-password.json
@@ -8,7 +8,7 @@
     "value": [
       {
         "rel": ["create-form"],
-        "name": "reenroll-custom-authenticator",
+        "name": "reenroll-custom-password-expiry",
         "relatesTo": ["$.currentAuthenticator"],
         "href": "https://sun.okta1.com/idp/idx/challenge/answer",
         "method": "POST",
