+++ src/v2/ion/ViewClassNamesFactory.js
@@ -59,7 +59,7 @@ const FORMNAME_CLASSNAME_MAPPINGS = {
   [FORMS.REENROLL_AUTHENTICATOR]: {
     [AUTHENTICATOR_KEY.PASSWORD]: 'password-expired'
   },
-  [FORMS.REENROLL_CUSTOM_AUTHENTICATOR]: {
+  [FORMS.REENROLL_CUSTOM_PASSWORD_EXPIRY]: {
     [AUTHENTICATOR_KEY.PASSWORD]: 'custom-password-expired'
   },
